Refactor `PreviewDraft`
Move all components to separate files to maybe help with readability
