+++ src/components/PreviewDraft/PreviewDraftLightboxV2.tsx
@@ -6,53 +6,24 @@
  *
  */
 
-import { useFormikContext } from "formik";
-import { ReactElement, useMemo, useState, ElementType } from "react";
-import { useTranslation } from "react-i18next";
+import { ReactElement, ElementType } from "react";
 import styled from "@emotion/styled";
 import { spacing } from "@ndla/core";
 import { ModalCloseButton, ModalHeader, ModalSizeType, Modal, ModalTrigger, ModalContent } from "@ndla/modal";
-import { IConcept } from "@ndla/types-backend/concept-api";
-import { IArticle } from "@ndla/types-backend/draft-api";
-import { OneColumn } from "@ndla/ui";
-import PreviewConceptComponent from "./PreviewConcept";
-import PreviewDraft from "./PreviewDraft";
-import { learningResourceFormTypeToDraftApiType } from "../../containers/ArticlePage/articleTransformers";
-import { ConceptFormValues } from "../../containers/ConceptPage/conceptInterfaces";
-import { conceptFormTypeToApiType } from "../../containers/ConceptPage/conceptTransformers";
-import { LearningResourceFormType } from "../../containers/FormikForm/articleFormHooks";
-import { useConcept } from "../../modules/concept/conceptQueries";
-import { useLicenses } from "../../modules/draft/draftQueries";
+import { ConceptPreviewProps, PreviewConcept } from "./PreviewConcept";
+import { CompareConceptPreviewProps, PreviewConceptCompare } from "./PreviewConceptCompare";
+import { MarkupPreviewProps, PreviewMarkup } from "./PreviewMarkup";
+import { PreviewVersion, VersionPreviewProps } from "./PreviewVersion";
 
-interface BaseProps {
+export interface PreviewBaseProps {
   type: "markup" | "version" | "conceptCompare" | "concept";
   language: string;
   activateButton?: ReactElement;
 }
 
-interface MarkupPreviewProps extends BaseProps {
-  type: "markup";
-  article: IArticle;
-}
-
-interface VersionPreviewProps extends BaseProps {
-  type: "version";
-  article: IArticle;
-  customTitle?: string;
-}
-
-interface CompareConceptPreviewProps extends BaseProps {
-  type: "conceptCompare";
-  concept: IConcept;
-}
-
-interface ConceptPreviewProps extends BaseProps {
-  type: "concept";
-}
-
 type Props = MarkupPreviewProps | VersionPreviewProps | CompareConceptPreviewProps | ConceptPreviewProps;
 
-const StyledPreviewWrapper = styled.div`
+export const StyledPreviewWrapper = styled.div`
   width: 100%;
   max-width: 100%;
   display: inline-flex;
@@ -72,23 +43,7 @@ const StyledPreviewWrapper = styled.div`
   }
 `;
 
-const PreviewMarkup = ({ article, language }: MarkupPreviewProps) => {
-  const { t } = useTranslation();
-  return (
-    <StyledPreviewWrapper>
-      <OneColumn>
-        <PreviewDraft
-          type="article"
-          draft={article}
-          language={language}
-          label={t("form.previewProductionArticle.article")}
-        />
-      </OneColumn>
-    </StyledPreviewWrapper>
-  );
-};
-
-const TwoArticleWrapper = styled(StyledPreviewWrapper)`
+export const TwoArticleWrapper = styled(StyledPreviewWrapper)`
   > div {
     margin: 0 2.5%;
     width: 40%;
@@ -101,120 +56,6 @@ const TwoArticleWrapper = styled(StyledPreviewWrapper)`
     }
   }
 `;
-
-const PreviewVersion = ({ article, language, customTitle }: VersionPreviewProps) => {
-  const { t } = useTranslation();
-  const { values, initialValues } = useFormikContext<LearningResourceFormType>();
-  const { data: licenses = [] } = useLicenses();
-  const formArticle = useMemo(() => {
-    const apiType = learningResourceFormTypeToDraftApiType(values, initialValues, licenses);
-    return {
-      id: article.id,
-      articleType: article.articleType,
-      title: apiType.title ?? "",
-      content: apiType.content ?? "",
-      introduction: apiType.introduction ?? "",
-      visualElement: apiType.visualElement,
-      published: apiType.published,
-      copyright: apiType.copyright,
-    };
-  }, [values, initialValues, licenses, article.id, article.articleType]);
-
-  return (
-    <TwoArticleWrapper>
-      <div>
-        <div className="u-10/12 u-push-1/12">
-          <h2>{t("form.previewProductionArticle.current")}</h2>
-        </div>
-        <PreviewDraft type="formArticle" draft={formArticle} language={language} label={article.articleType} />
-      </div>
-      <div>
-        <div className="u-10/12 u-push-1/12">
-          <h2>
-            {customTitle ??
-              t("form.previewProductionArticle.version", {
-                revision: article.revision,
-              })}
-          </h2>
-        </div>
-        <PreviewDraft type="article" draft={article} language={language} label={article.articleType} />
-      </div>
-    </TwoArticleWrapper>
-  );
-};
-
-const PreviewTitleWrapper = styled.div`
-  height: 90px;
-`;
-
-const ConceptWrapper = styled.div`
-  padding: 0 ${spacing.normal} ${spacing.normal} ${spacing.normal};
-`;
-
-const PreviewHeading = styled.h2`
-  margin: 0;
-`;
-
-const PreviewConceptCompare = ({ concept, language }: CompareConceptPreviewProps) => {
-  const [previewLanguage, setPreviewLanguage] = useState<string>(
-    concept.supportedLanguages.find((l) => l !== language) ?? concept.supportedLanguages[0]!,
-  );
-  const apiConcept = useConcept({ id: concept.id, language: previewLanguage });
-  const { data: licenses } = useLicenses({ placeholderData: [] });
-  const { t } = useTranslation();
-  const { values } = useFormikContext<ConceptFormValues>();
-  const formConcept = useMemo(
-    () => conceptFormTypeToApiType(values, licenses!, values.conceptType, concept.updatedBy),
-    [values, licenses, concept.updatedBy],
-  );
-  return (
-    <TwoArticleWrapper>
-      <ConceptWrapper>
-        <PreviewTitleWrapper>
-          <PreviewHeading>
-            {t("form.previewLanguageArticle.title", {
-              language: t(`languages.${language}`).toLowerCase(),
-            })}
-          </PreviewHeading>
-        </PreviewTitleWrapper>
-        <PreviewConceptComponent concept={formConcept} language={language} />
-      </ConceptWrapper>
-      <ConceptWrapper>
-        <PreviewTitleWrapper>
-          <PreviewHeading>
-            {t("form.previewLanguageArticle.title", {
-              language: t(`languages.${previewLanguage}`).toLowerCase(),
-            })}
-          </PreviewHeading>
-          <select onChange={(evt) => setPreviewLanguage(evt.target.value)} value={previewLanguage}>
-            {concept.supportedLanguages.map((language) => (
-              <option key={language} value={language}>
-                {t(`languages.${language}`)}
-              </option>
-            ))}
-          </select>
-        </PreviewTitleWrapper>
-        {apiConcept.data && <PreviewConceptComponent concept={apiConcept.data} language={previewLanguage} />}
-      </ConceptWrapper>
-    </TwoArticleWrapper>
-  );
-};
-
-const PreviewConcept = ({ language }: ConceptPreviewProps) => {
-  const { data: licenses } = useLicenses({ placeholderData: [] });
-  const { values } = useFormikContext<ConceptFormValues>();
-
-  const formConcept = useMemo(
-    () => conceptFormTypeToApiType(values, licenses!, values.conceptType),
-    [values, licenses],
-  );
-
-  return (
-    <ConceptWrapper>
-      <PreviewConceptComponent concept={formConcept} language={language} />
-    </ConceptWrapper>
-  );
-};
 
 const components: Record<Props["type"], ElementType> = {
   markup: PreviewMarkup,
